quantum_dashboard.py
#!/usr/bin/env python3
from flask import Flask, render_template_string
import os
from datetime import datetime

app = Flask(__name__)

HTML_TEMPLATE = """
<!DOCTYPE html>
<html>
<head>
    <title>Quantum Trader Dashboard</title>
    <meta charset="utf-8">
    <style>
        body { 
            font-family: 'Arial', sans-serif; 
            margin: 0; 
            padding: 20px; 
            background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
            color: white;
            min-height: 100vh;
        }
        .container {
            max-width: 1200px;
            margin: 0 auto;
        }
        .header {
            text-align: center;
            margin-bottom: 30px;
        }
        .cards {
            display: grid;
            grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
            gap: 20px;
            margin-bottom: 30px;
        }
        .card {
            background: rgba(255, 255, 255, 0.1);
            padding: 20px;
            border-radius: 15px;
            backdrop-filter: blur(10px);
            border: 1px solid rgba(255, 255, 255, 0.2);
        }
        .card h3 {
            margin-top: 0;
            color: #ffd700;
        }
        .status-active {
            color: #00ff00;
            font-weight: bold;
        }
        .log-container {
            background: rgba(0, 0, 0, 0.3);
            padding: 15px;
            border-radius: 10px;
            max-height: 400px;
            overflow-y: auto;
            font-family: 'Courier New', monospace;
            font-size: 12px;
        }
        .refresh-btn {
            background: #ffd700;
            color: #333;
            border: none;
            padding: 10px 20px;
            border-radius: 5px;
            cursor: pointer;
            font-weight: bold;
            margin: 10px;
        }
    </style>
</head>
<body>
    <div class="container">
        <div class="header">
            <h1>🚀 Quantum Trader Dashboard</h1>
            <p>Monitoraggio Sistema di Trading in Tempo Reale</p>
        </div>

        <div class="cards">
            <div class="card">
                <h3>💰 Portfolio</h3>
                <p><strong>XRPUSDT:</strong> 762.06 units</p>
                <p><strong>Balance:</strong> $9,300.00</p>
                <p><strong>Totale:</strong> ${{ portfolio_value }}</p>
                <p><strong>Trade Count:</strong> 3</p>
            </div>

            <div class="card">
                <h3>📊 Sistema</h3>
                <p><strong>Stato:</strong> <span class="status-active">ATTIVO</span></p>
                <p><strong>Ultimo Aggiornamento:</strong> {{ timestamp }}</p>
                <p><strong>Ultimo Ciclo:</strong> {{ last_cycle }}</p>
                <p><strong>Heartbeat:</strong> ✅ Funzionante</p>
            </div>

            <div class="card">
                <h3>🎯 Trading</h3>
                <p><strong>XRP Status:</strong> 🔒 BLOCCATO</p>
                <p><strong>Min Confluence:</strong> 2.6</p>
                <p><strong>Current Score:</strong> ~2.59</p>
                <p><strong>Signal:</strong> HOLD</p>
            </div>
        </div>

        <div class="card">
            <h3>📝 Logs Recenti</h3>
            <div class="log-container">
                {% for log in recent_logs %}
                    <div>{{ log }}</div>
                {% endfor %}
            </div>
        </div>

        <div style="text-align: center; margin-top: 20px;">
            <button class="refresh-btn" onclick="window.location.reload()">🔄 Aggiorna</button>
            <button class="refresh-btn" onclick="window.open('http://localhost:8000', '_blank')">📊 Nuova Finestra</button>
        </div>
    </div>
</body>
</html>
"""

def get_portfolio_value():
    try:
        with open('production.log', 'r') as f:
            for line in reversed(f.readlines()):
                if 'Portfolio:' in line:
                    return line.split('Portfolio: $')[-1].split()[0]
    except:
        pass
    return "11,340.03"

def get_last_cycle():
    try:
        with open('production.log', 'r') as f:
            for line in reversed(f.readlines()):
                if 'CICLO #' in line:
                    return line.split('CICLO #')[-1].split()[0]
    except:
        pass
    return "1"

@app.route('/')
def dashboard():
    try:
        with open('production.log', 'r') as f:
            logs = f.readlines()[-15:]
    except:
        logs = ["No logs available"]
    
    return render_template_string(HTML_TEMPLATE,
        timestamp=datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
        portfolio_value=get_portfolio_value(),
        last_cycle=f"#{get_last_cycle()}",
        recent_logs=logs)

if __name__ == '__main__':
    print("🚀 Quantum Dashboard avviato!")
    print("📍 APRI IL BROWSER: http://localhost:8000")
    print("🛑 Premi CTRL+C per fermare il dashboard")
    app.run(host='0.0.0.0', port=8000, debug=False)
