requirements.txt
python-binance>=1.0.19
streamlit>=1.28.0
pandas>=2.0.3
numpy>=1.24.3
requests>=2.31.0
plotly>=5.15.0
python-dotenv>=1.0.0
websocket-client>=1.6.3
ta-lib>=0.6.8
